requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
dbt-core==1.7.14
dbt-postgres==1.7.14
sqlalchemy==2.0.30
pyarrow==16.0.0
